fix(paths): record shared path package health budgets
